refactor: keep one copy of three functions duplicated across packages
Three helpers existed twice, byte for byte, in packages that already depend
on one another:

- `dataURLToBlob` in @ripl/dom and @ripl/terminal -> @ripl/core
- `computeTriangleNormal` in @ripl/3d and @ripl/webgpu -> vec3TriangleNormal
  in @ripl/3d's vector math, which webgpu already depends on
- `resolveInteraction` and its ResolvedInteraction shape in @ripl/dom and
  @ripl/3d -> @ripl/core, beside the NavigatorInteractionOption it resolves

No package gains a dependency; each move rides an existing edge.

vec3TriangleNormal keeps the flat scalar form rather than composing
vec3Sub/vec3Cross/vec3Normalize: it runs per triangle, and it wants the up
vector for a degenerate face where vec3Normalize yields zero.
