+++ packages/core/src/context/index.ts
@@ -1,3 +1,4 @@
+export * from './export';
 export * from './path';
 export * from './text';
 export * from './transform';
